% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARHT_one_sample.R
\name{ARHT_OnePop}
\alias{ARHT_OnePop}
\title{Adaptable Regularized Hotelling's T^2 test for one-sample testing problem}
\usage{
ARHT_OnePop(X, mu_0 = NULL, prob_alternative_prior = list(c(1, 0, 0), c(0,
  1, 0), c(0, 0, 1)), Type1error_calib = c("cube_root", "sqrt", "chi_sq",
  "none"), lambda_range = NULL, lambda_net_density = 2000,
  bootstrap_sample = NULL, bs_size = 1e+06)
}
\arguments{
\item{X}{n-by-p observation matrix with  numeric column variables.}

\item{mu_0}{the null hypothesis to be tested, default value is the 0 vector.}

\item{prob_alternative_prior}{an umempty list with each field a numeric vector of length 3 with sum 1.
Default value is list(c(1,0,0), c(0,1,0), c(0,0,1)).
Each field of the list represents a probabilistic prior models specified by weights
of $\Sigma^0$, $\Sigma^1$, $\Sigma^2$, etc.}

\item{Type1error_calib}{the method to calibrate Type 1 error rate of the test. Four values are allowed,
cube_root (default and recommended, cube-root transformation),
sqrt (square-root transformation),
chi_sq (chi-square approximation),
none (no calibration).}

\item{lambda_range}{a vector of two elements indicates the lower and upper bound of candidate lambda's;
If null, recommended choices will be calculated.}

\item{lambda_net_density}{postive numeric, default 2000. The number of grid points within the range of lambda's.
The grid is progressively coarser.}

\item{bs_size}{a numeric scalar, default value 1e6, only available when more than one prior models are
specified in prob_alternative_prior; control the size of bootstrap sample used to approximate p-values.}
}
\value{
Tstat: a vector of the same length of prob_alternative_prior.
}
\description{
Adaptable Regularized Hotelling's T^2 test for one-sample testing problem
}
