% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments_PSD.R
\name{moments_PSD}
\alias{moments_PSD}
\title{A function to calculate consistent estimators of the moments of the limiting spectral distribution of the population covariance matrix
given the spectral of the sample covariance matrix}
\usage{
moments_PSD(eigenvalues, n, mom_degree)
}
\arguments{
\item{n}{degree of freedom}

\item{emp_eig}{eigenvalues of the sample covariance matrix}

\item{degree}{the maximum order of the moments to calculate}
}
\value{
The first \code{mom_degree} number order of moments of the limiting spectral distribution of the population covariance matrix
ON ESTIMATION OF THE POPULATION SPECTRAL DISTRIBUTIONFROM A HIGH-DIMENSIONAL SAMPLE COVARIANCE MATRIX
Australian & New Zealand Journal of Statistics.
}
\description{
A function to calculate consistent estimators of the moments of the limiting spectral distribution of the population covariance matrix
given the spectral of the sample covariance matrix
}
\keyword{estimators}
\keyword{moments}
\keyword{population}
\keyword{spectral}
