% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3chisq.R
\name{r3chisq}
\alias{r3chisq}
\title{3-variate positively correlated chi-squared distribution generation when the degrees of freedom are large}
\usage{
r3chisq(size, df, correlation_mat)
}
\arguments{
\item{size}{sample size}

\item{df}{the degree of freedoms \eqn{(d_1, d_2, d_3)} (non-negative, but can be non-integer, \code{ceiling(df)} is employed if non-integer)}

\item{correlation_mat}{the desired correlation matrix; negative elements will be set to 0}
}
\value{
A list of
\itemize{
\item{sample}: a n-by-3 matrix contains the generated sample of size n
\item{approx_cov}: the true covariance matrix of the returned sample}
Example: cor_mat = matrix(c(1, 1/6,2/3, 1/6, 1, 2/3, 2/3, 2/3, 1),3,3)
example1 = r3chisq(size = 10000, df =c(80,90,100), correlation_mat = cor_mat )
cov(example1$sample) - example1$approx_cov
cov2cor(example1$approx_cov) - cor_mat
}
\description{
Generate samples approximately from three positively correlated chi-square random variables \eqn{(\chi^2(d_1), \chi^2(d_2), \chi^2(d_3))}
 when the degrees of freedoms \eqn{(d_1, d_2, d_3)} are large
}
\details{
It is generally hard to sample from \eqn{(\chi^2(d_1), \chi^2(d_2), \chi^2(d_3))} with a designed correlation matrix. In the alogrithm, we approximate
the random vector by \eqn{(z^T Q_1 z, z^T Q_2 z, z^T Q_3 z)} where \eqn{z} is a standard norm random vector and \eqn{Q_1,Q_2,Q_3} are diagonal matrices
with diagonal elements 1's and 0's. We approximate the designed positive correlations by carafully selecting common locations of 1's on the diagonals.
The generated sample may have slightly different marginal degrees of freedoms from the inputted \code{df}, also covariances.
}
