% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARHT.R
\name{ARHT}
\alias{ARHT}
\title{Adaptable Regularized Hotelling's T^2 test for one-sample testing problem}
\usage{
ARHT(X, Y = NULL, mu_0 = NULL, prob_alternative_prior = list(c(1, 0, 0),
  c(0, 1, 0), c(0, 0, 1)), Type1error_calib = c("cube_root", "sqrt", "chi_sq",
  "none"), lambda_range = NULL, nlambda = 2000, bs_size = 1e+06)
}
\arguments{
\item{X}{n-by-p observation matrix with  numeric column variables.}

\item{mu_0}{the null hypothesis to be tested, default value is the 0 vector.}

\item{prob_alternative_prior}{an umempty list with each field a numeric vector of length 3 with sum 1.
Default value is list(c(1,0,0), c(0,1,0), c(0,0,1)).
Each field of the list represents a probabilistic prior models specified by weights
of $\Sigma^0$, $\Sigma^1$, $\Sigma^2$, etc.}

\item{Type1error_calib}{the method to calibrate Type 1 error rate of the test. Four values are allowed,
cube_root (default, cube-root transformation),
sqrt (square-root transformation),
chi_sq (chi-square approximation),
none (no calibration).}

\item{lambda_range}{Optional user-supplied lambda sequence; default is NULL, and ARHT chooses
its own sequence}

\item{nlambda}{Optional user-supplied number of lambda's in grid search; Default is 2000; Must be postive numeric. The grid is progressively coarser.}

\item{bs_size}{a numeric scalar, default value 1e6, only available when more than one prior models are
specified in prob_alternative_prior; control the size of bootstrap sample used to approximate p-values.}
}
\value{
Tstat: a vector of the same length of prob_alternative_prior.
}
\description{
Adaptable Regularized Hotelling's T^2 test for one-sample testing problem
}
