% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_proj_2samp.R
\name{eigen_proj_2samp}
\alias{eigen_proj_2samp}
\title{An internal function for ARHT}
\usage{
eigen_proj_2samp(X, Y, mu_0, lower_lambda = NULL)
}
\arguments{
\item{X}{X matrix as in ARHT}

\item{Y}{Y matrix as in ARHT}

\item{mu_0}{mu_0 as in ARHT}

\item{lower_lambda}{the lower bound of lambda sequence, that is lambda_range[1] as in ARHT if specified; otherwise NULL}
}
\value{
A list of
\itemize{\item{\code{n}}: degree of freedom, that is the sum of two sample sizes minus 2
         \item{\code{pos_eig_val}}: the eigenvalue of the sample covariance matrix
         \item{\code{proj_shift}}: the scaled projection of the distance between the sample mean and mu_0; see description
         \item{\code{lower_bound}}: If lower_lambda specified, lower_bound returns its value; if not, returns the generated lower bound of lambda sequence
}
}
\description{
The function takes X, Y and mu_0 as input, and returns positives eigenvalues of the pooled sample covariance matrix,
             and the scaled projection of the distance between the difference of the sample means and mu_0, that is,
             \deqn{\frac{N_1N_2}{N1+N2} P^T (\bar{X}- \bar{Y} - mu_0)}
             where $P$ is the eigenvector matrix of the pooled sample covariance matrix. The function is designed to handle the situation when
             the algorithm of singular value decomposition of the sample covariance fails to converge, by adding ridge-term to
             stablize the algorithm.
}
\keyword{internal}
