% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments_PSD.R
\name{moments_PSD}
\alias{moments_PSD}
\title{A function to calculate a consistent estimator of $\frac{1}{p}tr(\Sigma^m)$ for any positive integer m,
given the spectral of the sample covariance matrix $S_n$.
Reference Lemma 1 of Bai, Chen, Yao (2010),
ON ESTIMATION OF THE POPULATION SPECTRAL DISTRIBUTIONFROM A HIGH-DIMENSIONAL SAMPLE COVARIANCE MATRIX
Australian & New Zealand Journal of Statistics.}
\usage{
moments_PSD(eigenvalues, n, degree)
}
\arguments{
\item{degree}{the order of the moment.}

\item{emp_eig}{eigenvalues of S_n.}

\item{gamma}{p/n.}
}
\value{
The first \code{degree} number order of moments of the distribution spectral of Sigma.
}
\description{
A function to calculate a consistent estimator of $\frac{1}{p}tr(\Sigma^m)$ for any positive integer m,
given the spectral of the sample covariance matrix $S_n$.
Reference Lemma 1 of Bai, Chen, Yao (2010),
ON ESTIMATION OF THE POPULATION SPECTRAL DISTRIBUTIONFROM A HIGH-DIMENSIONAL SAMPLE COVARIANCE MATRIX
Australian & New Zealand Journal of Statistics.
}
